backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary
alembic>=1.10.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic>=2.0.0
pydantic-settings
httpx
pytest
pytest-asyncio
redis
celery
pillow
jinja2
email-validator 